fix: rebalance loss weights — depth_gt dominated at 66%, warp warmup 8x too slow
depth_gt (SiLog on sparse LiDAR) weight 0.5 produced 10.28/15.55 = 66% of
total loss, hitting a floor quickly and masking all other signal. Reduced to
0.05 to make it a gentle regularizer (~17%).

Warp warmup with 2 full epochs on 177k-sample T4 dataset meant 44k warmup
steps — warp wouldn't reach 100% until well past useful training window.
Changed to 0.25 epochs (5.5k steps), matching original PG-Occ's ~3.5k step
warmup on nuScenes. Also fixed type hint from int to float.

New loss balance: depth_foundation 78%, depth_gt 17%, warp 2% (ramping to
~25% at full strength), OV 3%.

+++ models/pgocc/pgocc.py
@@ -55,7 +55,7 @@ class PGOccLightning(pl.LightningModule):
         ov_dim: int = 768,
         # Losses
         loss_weights: dict = None,
-        warp_warmup_epochs: int = 2,
+        warp_warmup_epochs: float = 2,
         # Masking
         ego_car_mask_dir: str = "",
         ego_car_mask_map: dict = None,
